+++ src/focal.c
@@ -25,9 +25,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#ifdef _OPENMP
-#include <omp.h>
-#endif
+#include "vec_omp.h"
 
 #ifndef M_PI
 #define M_PI 3.14159265358979323846